Integer ->Long(erp.safe)

+++ erp-logics/src/main/java/lsfusion/erp/region/by/machinery/cashregister/fiscalepson/FiscalEpsonPrintReceiptActionProperty.java
@@ -149,7 +149,7 @@ public class FiscalEpsonPrintReceiptActionProperty extends ScriptingActionProper
                                                 sumGiftCard == null ? null : sumGiftCard.abs(), cashier,
                                                 isReturn ? receiptReturnItemList : receiptSaleItemList)));
                         if (result instanceof Integer) {
-                            findProperty("number[Receipt]").change(result, context, receiptObject);
+                            findProperty("number[Receipt]").change((Integer)result, context, receiptObject);
                             if (context.apply())
                                 findAction("createCurrentReceipt[]").execute(context);
                             else
